// gRPCServer/protos-out/protos/message_pb.d.ts
  export type AsObject = {
  }
// ...
export   class WalletInfoRequest extends jspb.Message {
  getAddress(): string;
  setAddress(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): WalletInfoRequest.AsObject;
  static toObject(includeInstance: boolean, msg: WalletInfoRequest): WalletInfoRequest.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: WalletInfoRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): WalletInfoRequest;
  static deserializeBinaryFromReader(message: WalletInfoRequest, reader: jspb.BinaryReader): WalletInfoRequest;
}

export namespace WalletInfoRequest {
  export type AsObject = {
    address: string,
  }
}

export class WalletInfoResponse extends jspb.Message {
  getTotal(): number;
  setTotal(value: number): void;

  getAvailable(): number;
  setAvailable(value: number): void;

  clearTransactionList(): void;
  getTransactionList(): Array<Transaction>;
  setTransactionList(value: Array<Transaction>): void;
  addTransaction(value?: Transaction, index?: number): Transaction;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): WalletInfoResponse.AsObject;
  static toObject(includeInstance: boolean, msg: WalletInfoResponse): WalletInfoResponse.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: WalletInfoResponse, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): WalletInfoResponse;
  static deserializeBinaryFromReader(message: WalletInfoResponse, reader: jspb.BinaryReader): WalletInfoResponse;
}

export namespace WalletInfoResponse {
  export type AsObject = {
    total: number,
    available: number,
    transactionList: Array<Transaction.AsObject>,
  }
}

export class Transaction extends jspb.Message {
  getToAddress(): string;
  setToAddress(value: string): void;

  getPoints(): number;
  setPoints(value: number): void;

  getMetadata(): string;
  setMetadata(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Transaction.AsObject;
  static toObject(includeInstance: boolean, msg: Transaction): Transaction.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: Transaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Transaction;
  static deserializeBinaryFromReader(message: Transaction, reader: jspb.BinaryReader): Transaction;
}

export namespace Transaction {
  export type AsObject = {
    toAddress: string,
    points: number,
    metadata: string,
  }
}